// src/payment-methods/enums/payment-method-error-code.enum.ts
export enum PaymentMethodErrorCode {
  PAYMENT_METHOD_NOT_FOUND = 'PAYMENT_METHOD_NOT_FOUND',
  PAYMENT_METHOD_FORBIDDEN = 'PAYMENT_METHOD_FORBIDDEN',
  PAYMENT_METHOD_SYSTEM = 'PAYMENT_METHOD_SYSTEM',
  PAYMENT_METHOD_DUPLICATE = 'PAYMENT_METHOD_DUPLICATE',
  PAYMENT_METHOD_ALREADY_DELETED = 'PAYMENT_METHOD_ALREADY_DELETED',
  PAYMENT_METHOD_NOT_DELETED     = 'PAYMENT_METHOD_NOT_DELETED',
  PAYMENT_METHOD_FAILED = 'PAYMENT_METHOD_FAILED',
}